ATA/RAG_Retrieval/requirements.txt
# Core Dependencies (CUDA-enabled)
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Vector Database
chromadb>=0.4.15
faiss-gpu>=1.7.4  # For GPU support, or faiss-cpu for CPU-only

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Translation & NLP
sacrebleu>=2.3.1
nltk>=3.8.1

# API & UI
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Model optimization for low VRAM
optimum>=1.16.0
peft>=0.7.0
